requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
mediapipe>=0.8.10
tensorflow>=2.5.0
keras>=2.5.0
pandas>=1.3.0
Pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
face-recognition>=1.3.0
dlib>=19.22.0
imutils>=0.5.4
openpyxl>=3.0.7
